Improve rotary encoder handling
Now rotary encoder is readout leveraging the hardware pulse counter
embedded in the ESP32 platform, using Zephyr's drivers.
No more false reads, regardless of the polling rate.
getChSelector return value was moved to unsigned 8 bit and the
returned value of the pulse encoder in degrees is translated to modulo
255.

TG-553