Add function to register extra Info menu entries
Added the option to add extra Info menu entries, extra entries take the
form of key-value pairs where the key is a constant string and the value
is the return value of a callback. This allows platform-dependent code
to inject new entries without filling the UI of platform-dependent code.
Note that the string returned by the callback is allocated dinamically,
thus we need malloc support and care has to be taken by the Info menu
rendering code to free the returned string.

TG-553