Enable Tx capability, band detection
Enabled Tx capability on T-TWR Plus.
Since Tx requires the SA8x8 to disable the Rx stage, we check that at
least firmware v.1.1.0.r14 is present on the baseband, otherwise we
don't initialize the radio.
Band detection is added, that queries the SA8x8 for the supported band.

TG-553